Refactor Wordmove::Logger methods
Now they wrap around ::Logger's methods. This
approach should open the way to a cleaner approach
using loglevels.
For example when wordmove is called in debug mode
we should not infest the code with conditionals that
control if the log message will be printed, but we
should simply update the loglevel once on the logger
instance.

A refactor will follow. Not sure if within the current
branch